Bug fix for new numpy version

+++ pyworkflow/em/packages/xmipp3/protocol_particle_pick_consensus.py
@@ -116,7 +116,7 @@ class XmippProtConsensusPicking(ProtParticlePicking):
         n=0
         for coordinates in self.inputCoordinates:
             coordArray = np.asarray([x.getPosition() 
-                                     for x in coordinates.get().iterCoordinates(micId)])
+                                     for x in coordinates.get().iterCoordinates(micId)],dtype=np.float)
             coordArray *= Tm[n]/Tm[0]
             coords.append(coordArray)
             Ncoords += coordArray.shape[0]
